Move utils function outside
